requirements.txt
torch
torchvision
transformers
scikit-learn
pillow
tqdm
matplotlib
faiss-cpu